+++ packages/cli/CHANGELOG.md
@@ -1,10 +1,23 @@
 # @intent-driven/cli
 
-## 1.0.23
+## 1.0.24
 
 ### Patch Changes
 
 - Updated dependencies [2e02b73]
+  - @intent-driven/core@0.23.0
+
+## 1.0.23
+
+### Patch Changes
+
+- Updated dependencies [eb2954d]
+- Updated dependencies [eb2954d]
+- Updated dependencies [eb2954d]
+- Updated dependencies [eb2954d]
+- Updated dependencies [eb2954d]
+- Updated dependencies [eb2954d]
+- Updated dependencies [eb2954d]
   - @intent-driven/core@0.22.0
 
 ## 1.0.22